requirements.txt
django
psycopg2
gunicorn
dj_database_url
python-dotenv
Pillow